refactor(inspector): keep only the diff-content prefetch drop
Profiling on a 594-file workspace showed neither the virtualized list
nor the content-visibility pseudo-virt buys anything: virt is actually
slightly worse during fast scrolls because reconcile competes with
paint, and pseudo-virt is indistinguishable from baseline. The real
CPU win was always from dropping the eager workspaceChanges prefetch.
Revert both rendering changes; keep just the prefetch removal and
clean up the surfaced dead code on both sides (also drops the long-
unused listEditorFilesWithContent path that no longer had any caller).
