+++ src-tauri/src/commands/editor_commands.rs
@@ -34,25 +34,10 @@ pub async fn list_workspace_files(
 }
 
 #[tauri::command]
-pub async fn list_editor_files_with_content(
-    workspace_root_path: String,
-) -> CmdResult<editor_files::EditorFilesWithContentResponse> {
-    run_blocking(move || editor_files::list_editor_files_with_content(&workspace_root_path)).await
-}
-
-#[tauri::command]
 pub async fn list_workspace_changes(
     workspace_root_path: String,
 ) -> CmdResult<Vec<editor_files::EditorFileListItem>> {
     run_blocking(move || editor_files::list_workspace_changes(&workspace_root_path)).await
-}
-
-#[tauri::command]
-pub async fn list_workspace_changes_with_content(
-    workspace_root_path: String,
-) -> CmdResult<editor_files::EditorFilesWithContentResponse> {
-    run_blocking(move || editor_files::list_workspace_changes_with_content(&workspace_root_path))
-        .await
 }
 
 #[tauri::command]
