+++ src-tauri/src/lib.rs
@@ -369,10 +369,8 @@ pub fn run() {
             commands::workspace_commands::pin_workspace,
             commands::workspace_commands::unpin_workspace,
             commands::editor_commands::list_editor_files,
-            commands::editor_commands::list_editor_files_with_content,
             commands::editor_commands::list_workspace_files,
             commands::editor_commands::list_workspace_changes,
-            commands::editor_commands::list_workspace_changes_with_content,
             commands::editor_commands::discard_workspace_file,
             commands::editor_commands::stage_workspace_file,
             commands::editor_commands::unstage_workspace_file,
