+++ src-tauri/src/workspace/files/editor.rs
@@ -11,8 +11,7 @@ use super::{
         editor_file_sort_key, metadata_mtime_ms, resolve_allowed_path,
     },
     types::{
-        EditorFileListItem, EditorFilePrefetchItem, EditorFileReadResponse, EditorFileStatResponse,
-        EditorFileWriteResponse, EditorFilesWithContentResponse,
+        EditorFileListItem, EditorFileReadResponse, EditorFileStatResponse, EditorFileWriteResponse,
     },
 };
 use crate::{
@@ -21,7 +20,6 @@ use crate::{
 };
 
 const MAX_EDITOR_FILE_ITEMS: usize = 24;
-const MAX_PREFETCH_BYTES: u64 = 1_048_576;
 
 /// Read a file at a given git ref. Returns `None` when the path doesn't
 /// exist in that ref, or when the workspace itself has vanished (e.g. the
@@ -190,15 +188,6 @@ pub fn list_workspace_files(workspace_root_path: &str) -> Result<Vec<EditorFileL
     Ok(build_list_items(&workspace_root, discovered_files))
 }
 
-pub fn list_editor_files_with_content(
-    workspace_root_path: &str,
-) -> Result<EditorFilesWithContentResponse> {
-    let items = list_editor_files(workspace_root_path)?;
-    let prefetched = prefetch_items(&items, true);
-
-    Ok(EditorFilesWithContentResponse { items, prefetched })
-}
-
 /// Best-effort variant for read-only listers. Returns `None` if the
 /// workspace directory has vanished (deleted externally, archived, etc.)
 /// so callers surface an empty list instead of a red toast. Real errors
@@ -244,29 +233,6 @@ fn build_list_items(
                 staged_status: None,
                 unstaged_status: None,
                 committed_status: None,
-            })
-        })
-        .collect()
-}
-
-fn prefetch_items(
-    items: &[EditorFileListItem],
-    include_deleted: bool,
-) -> Vec<EditorFilePrefetchItem> {
-    items
-        .iter()
-        .filter(|item| include_deleted || item.status != "D")
-        .filter_map(|item| {
-            let path = Path::new(&item.absolute_path);
-            let metadata = fs::metadata(path).ok()?;
-            if metadata.len() > MAX_PREFETCH_BYTES {
-                return None;
-            }
-            let bytes = fs::read(path).ok()?;
-            let content = String::from_utf8(bytes).ok()?;
-            Some(EditorFilePrefetchItem {
-                absolute_path: item.absolute_path.clone(),
-                content,
             })
         })
         .collect()
