+++ src-tauri/src/workspace/files/mod.rs
@@ -4,16 +4,14 @@ mod support;
 mod types;
 
 pub use changes::{
-    discard_workspace_file, list_workspace_changes, list_workspace_changes_with_content,
-    stage_workspace_file, unstage_workspace_file,
+    discard_workspace_file, list_workspace_changes, stage_workspace_file, unstage_workspace_file,
 };
 pub use editor::{
-    list_editor_files, list_editor_files_with_content, list_workspace_files, read_editor_file,
-    read_file_at_ref, stat_editor_file, write_editor_file,
+    list_editor_files, list_workspace_files, read_editor_file, read_file_at_ref, stat_editor_file,
+    write_editor_file,
 };
 pub use types::{
-    EditorFileListItem, EditorFilePrefetchItem, EditorFileReadResponse, EditorFileStatResponse,
-    EditorFileWriteResponse, EditorFilesWithContentResponse,
+    EditorFileListItem, EditorFileReadResponse, EditorFileStatResponse, EditorFileWriteResponse,
 };
 
 #[cfg(test)]
