+++ src/features/inspector/sections/actions.tsx
@@ -51,9 +51,7 @@ import {
 	INSPECTOR_ACTIONS_BODY_VAR,
 	INSPECTOR_SECTION_HEADER_CLASS,
 	INSPECTOR_SECTION_HEADER_HEIGHT,
-	INSPECTOR_SECTION_HEIGHT_TRANSITION_CLASS,
 	INSPECTOR_SECTION_TITLE_CLASS,
-	INSPECTOR_SECTION_TOGGLE_ICON_CLASS,
 } from "../layout";
 
 interface GitStatusItem {
@@ -110,7 +108,6 @@ type ActionsSectionProps = {
 	workspaceRemote?: string | null;
 	sectionRef?: React.RefObject<HTMLElement | null>;
 	open: boolean;
-	resizing?: boolean;
 	onToggle: () => void;
 	bodyHeight: number;
 	onCommitAction?: (mode: WorkspaceCommitButtonMode) => Promise<void>;
@@ -163,7 +160,6 @@ export function ActionsSection({
 	workspaceRemote,
 	sectionRef,
 	open,
-	resizing = false,
 	onToggle,
 	bodyHeight,
 	onCommitAction,
@@ -331,9 +327,6 @@ export function ActionsSection({
 			aria-label="Inspector section Actions"
 			className={cn(
 				"flex min-h-0 shrink-0 flex-col overflow-hidden border-b border-border/60 bg-sidebar transition-colors",
-				resizing
-					? "transition-none"
-					: INSPECTOR_SECTION_HEIGHT_TRANSITION_CLASS,
 			)}
 			style={{
 				// Height via CSS var, written by mousemove during drag. Collapsed
@@ -361,11 +354,12 @@ export function ActionsSection({
 					className="shrink-0 text-muted-foreground hover:bg-accent/60 hover:text-foreground"
 				>
 					<ChevronDown
-						className={cn(
-							INSPECTOR_SECTION_TOGGLE_ICON_CLASS,
-							!open && "-rotate-90",
-						)}
+						className="size-3.5"
 						strokeWidth={1.9}
+						style={{
+							transform: open ? "rotate(0deg)" : "rotate(-90deg)",
+							transition: "none",
+						}}
 					/>
 				</Button>
 			</div>
