--- src/features/inspector/sections/changes/change-row.tsx
@@ -1,4 +1,0 @@
-export { ChangeFileRow } from "./file-row";
-export { ChangeFolderRow } from "./folder-row";
-export { ChangeGroupHeaderRow } from "./group-row";
-export { EmptyChangesRow, LoadingChangesRow } from "./placeholder-rows";
