+++ src/features/inspector/sections/changes/use-changes-state.ts
@@ -1,7 +1,8 @@
 // Toggle state for the four collapsible regions of the Changes section
 // (Changes header, Staged, Branch Diff) plus the two tree/flat view
-// toggles. Stable callbacks keep the derived changes model memoizable.
-import { useCallback, useState } from "react";
+// toggles. Plain boolean state with stable setters so the JSX consumer
+// can pass them straight into onClick handlers.
+import { useState } from "react";
 
 export type ChangesStateController = {
 	changesOpen: boolean;
@@ -22,26 +23,6 @@ export function useChangesState(): ChangesStateController {
 	const [changesOpen, setChangesOpen] = useState(true);
 	const [stagedOpen, setStagedOpen] = useState(true);
 	const [branchDiffOpen, setBranchDiffOpen] = useState(true);
-	const toggleChangesOpen = useCallback(
-		() => setChangesOpen((current) => !current),
-		[],
-	);
-	const toggleStagedOpen = useCallback(
-		() => setStagedOpen((current) => !current),
-		[],
-	);
-	const toggleBranchDiffOpen = useCallback(
-		() => setBranchDiffOpen((current) => !current),
-		[],
-	);
-	const toggleChangesTreeView = useCallback(
-		() => setChangesTreeView((current) => !current),
-		[],
-	);
-	const toggleBranchDiffTreeView = useCallback(
-		() => setBranchDiffTreeView((current) => !current),
-		[],
-	);
 
 	return {
 		changesOpen,
@@ -49,10 +30,10 @@ export function useChangesState(): ChangesStateController {
 		branchDiffOpen,
 		changesTreeView,
 		branchDiffTreeView,
-		toggleChangesOpen,
-		toggleStagedOpen,
-		toggleBranchDiffOpen,
-		toggleChangesTreeView,
-		toggleBranchDiffTreeView,
+		toggleChangesOpen: () => setChangesOpen((current) => !current),
+		toggleStagedOpen: () => setStagedOpen((current) => !current),
+		toggleBranchDiffOpen: () => setBranchDiffOpen((current) => !current),
+		toggleChangesTreeView: () => setChangesTreeView((v) => !v),
+		toggleBranchDiffTreeView: () => setBranchDiffTreeView((v) => !v),
 	};
 }
