+++ src/lib/api.ts
@@ -449,16 +449,6 @@ export type EditorFileStatResponse = {
 	size: number | null;
 };
 
-export type EditorFilePrefetchItem = {
-	absolutePath: string;
-	content: string;
-};
-
-export type EditorFilesWithContentResponse = {
-	items: InspectorFileItem[];
-	prefetched: EditorFilePrefetchItem[];
-};
-
 export type AppUpdateStage =
 	| "disabled"
 	| "idle"
@@ -1922,40 +1912,11 @@ export async function listWorkspaceFiles(
 
 export async function listWorkspaceChanges(
 	workspaceRootPath: string,
-): Promise<EditorFilesWithContentResponse> {
+): Promise<InspectorFileItem[]> {
 	try {
-		const items = await invoke<InspectorFileItem[]>("list_workspace_changes", {
+		return await invoke<InspectorFileItem[]>("list_workspace_changes", {
 			workspaceRootPath,
 		});
-		return { items, prefetched: [] };
-	} catch (error) {
-		throw new Error(
-			describeInvokeError(error, "Unable to list workspace changes."),
-		);
-	}
-}
-
-export async function listEditorFilesWithContent(
-	workspaceRootPath: string,
-): Promise<EditorFilesWithContentResponse> {
-	try {
-		return await invoke<EditorFilesWithContentResponse>(
-			"list_editor_files_with_content",
-			{ workspaceRootPath },
-		);
-	} catch (error) {
-		throw new Error(describeInvokeError(error, "Unable to list editor files."));
-	}
-}
-
-export async function listWorkspaceChangesWithContent(
-	workspaceRootPath: string,
-): Promise<EditorFilesWithContentResponse> {
-	try {
-		return await invoke<EditorFilesWithContentResponse>(
-			"list_workspace_changes_with_content",
-			{ workspaceRootPath },
-		);
 	} catch (error) {
 		throw new Error(
 			describeInvokeError(error, "Unable to list workspace changes."),
