+++ src/test/e2e-mocks/invoke-defaults.ts
@@ -33,7 +33,6 @@ export const defaultInvokeHandlers: Record<string, InvokeHandler> = {
 	list_remote_branches: () => [],
 	list_workspace_files: () => [],
 	list_workspace_changes: () => [],
-	list_workspace_changes_with_content: () => ({ items: [], prefetched: [] }),
 	list_slash_commands: () => [],
 	list_forge_labels: () => [],
 	list_inbox_kind_labels: () => [],
