+++ src/test/setup.ts
@@ -204,8 +204,6 @@ vi.mock("@tauri-apps/api/core", () => ({
 				return [];
 			case "list_workspace_changes":
 				return [];
-			case "list_workspace_changes_with_content":
-				return { items: [], prefetched: [] };
 			case "list_slash_commands":
 				return [];
 			case "list_workspace_linked_directories":
